Implementações/RNF-acuracia/modeloKNN/requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.5.0
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
asyncio==3.4.3
blinker==1.6.2
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.2.0
confection==0.1.3
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.7
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
Flask==2.3.3
Flask-Cors==4.0.0
fonttools==4.45.1
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
gensim==4.3.2
ibm-cloud-sdk-core==3.16.7
ibm-watson==7.0.1
idna==3.4
imbalanced-learn==0.11.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
ipykernel==6.27.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
langcodes==3.3.0
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
murmurhash==1.0.9
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.0
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.3.0
preshed==3.0.8
prompt-toolkit==3.0.41
psutil==5.9.6
pt-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.6.0/pt_core_news_lg-3.6.0-py3-none-any.whl#sha256=4a15fab02cd32ba7bc2256c73ad453eb9d307f1393d825d77cb296028422deb3
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pytest==7.4.2
pytest-env==1.0.1
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
smart-open==6.4.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
stack-data==0.6.3
thinc==8.1.12
threadpoolctl==3.2.0
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wasabi==1.1.2
wcwidth==0.2.12
websocket-client==1.6.1
Werkzeug==2.3.7
zipp==3.17.0
